recombine jobs*.py into jobs.py
